requirements.txt
moviepy==1.0.3
pydub==0.25.1
gtts==2.2.3
requests==2.26.0
google-generativeai==0.1.0